Switch State::views from set<View *> to vector<View *>.
set<View *> orders elements by their addresses, which will vary
unpredictably from process to process, particularly if the operating
system does address space layout randomization.

This may cause a performance regression, if removing empty views
happens frequently.  In that case, we can introduce a set<unsigned>
of the currently active view indices, in order to use the vector for
ordering and the set for fast iteration of active views.