Clean up simple_predictive_sample_observed.
I find function definitions much easier to read than reified dicts,
because it is not necessary to understand the control flow of
reification to understand what the function does; and the code
structure is now much clearer: some preparatory definitions and a loop
that uses them.

Retained the caching that those dicts provided explicitly as a cache
on the function I expect to be expensive.