Fix name and value of mu hyperparamter for continuous model test.
The test was written for what *happened* to be the `randomly' chosen
value, under the fixed seed 0, for mu, at the beginning of the test.
It appears to have been intended to be written for mu = 13, but that
doesn't match the first computed logp value.  And then the name `mu'
was mistyped as `m', a mistake which was masked because the hypers map
already had an entry correctly spelled `mu'.