Fix one-character typo in modified Bessel function i_1.
Either

    v *= -1;

or

    v = -v;

works better to negate v.

This does not affect anything in Crosscat because the von Mises
distribution relies only on nonnegative inputs to i_1, and although
Crosscat may pass negative inputs in during Newton-Raphson iteration
to compute a maximum-likelihood estimate for the von Mises kappa, if
we returned a negative root then that would be a bug anyway.