Define __STDC_CONSTANT_MACROS to make UINT64_C available.
This macro enables C99 stdint.h to work in C++98.  The need for this
was obviated by C++11, and the macro removed in C11, so newer glibc
stdint.h does not require __STDC_CONSTANT_MACROS even for C++98, but
defining the macro makes it easier to use this with slightly older
versions of glibc (such as the GCC 4.7 found in Debian Wheezy) or
other systems and is harmless otherwise.

Should fix Github issue #96.