Moves recording of diagnostics from LocalEngine.pyx to State.pyx
Two reasons for this commit:
(i) The current implementation of diagnostic recording fails when the user
specifies a `max_time` with `diagnostics_every_N`. This is beacuse _each invocation
of p_State.transition uses the same `max_time` inside the for loop over
`child_n_steps`, whereas we really need `max_time` to apply to the union of
calls to p_State.transition.

(ii) The progress bar is now compatible with diagnostic recording, which
occurs in the inner loop of State.transition.