Make sure that the column_dependencies map contains connected components.
NOTICE: Currently depends on py-tarjan, an implementation of Tarjan's algorithm
for strongly connected components from:

    https://github.com/bwesterb/py-tarjan/

We should decide what to do, if anything, about this dependency.

In the past, we had a map of the form

    {
        0: [1,2],
        2: [3],
    }

which is clearly not compatible with block transitions. Therefore, the
data structure passed into the State.cpp is now guaranteed to contain the
connected components implied by the ENSURE DEPENDENT constraints, e.g.

    {
        0 : [0,1,2,3],
        1 : [0,1,2,3],
        2 : [0,1,2,3],
        3 : [0,1,2,3],
    }

which ensures that block proposals happen correctly regardless of the feature
being transitioned.