Add utils for get_vector_num_blocks.
This function will be used to find the ``effective'' number of columns
in a CrossCat view, as well as an ``effective'' number of columns in a
CrossCat state.

Paltry test in cpp_code/tests/test_utils.cpp -- can't figure out a nice
way to build the data structures that will serve as the test cases.